Cleaning for manual import of custom tokens
